pyproject.toml
[project]
name = "ai-jdr"
version = "0.1.0"
description = "A simple AI app to build your own homebrew tabletop games"
authors = [
    {name = "Gabriel Hurtado"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "psycopg (>=3.2.6,<4.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "propelauth-fastapi (>=4.2.6,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
